+++ src/d/actor/d_a_npc_fairy.cpp
@@ -161,7 +161,7 @@ __ct__13daNpc_Fairy_cFPC26daNpcT_faceMotionAnmData_cPC22daNpcT_motionAnmData_cPC
 extern "C" void __dt__12J3DFrameCtrlFv();
 extern "C" void __dt__8cM3dGCylFv();
 extern "C" void __dt__8cM3dGAabFv();
-extern "C" s32 getEyeballMaterialNo__13daNpc_Fairy_cFv();
+extern "C" u16 getEyeballMaterialNo__13daNpc_Fairy_cFv();
 extern "C" s32 getHeadJointNo__13daNpc_Fairy_cFv();
 extern "C" s32 getNeckJointNo__13daNpc_Fairy_cFv();
 extern "C" bool getBackboneJointNo__13daNpc_Fairy_cFv();
@@ -2699,7 +2699,7 @@ extern "C" void __dt__8cM3dGAabFv() {
 }
 
 /* 809B9238-809B9240 007818 0008+00 1/0 0/0 0/0 .text getEyeballMaterialNo__13daNpc_Fairy_cFv */
-s32 daNpc_Fairy_c::getEyeballMaterialNo() {
+u16 daNpc_Fairy_c::getEyeballMaterialNo() {
     return true;
 }
